Implemented model sections, method-based line item creation, and `.previous()` (#3)
Did a lot of things at once in this PR. To summarise them:

- I added the concept of a `Section`. This is a data structure that
collects many line items together. Helps to organise models, and
namespaces all of the columns in the final model with this structure
alias. Also allows users to define section "outputs" to help make the
model more readable.
- Added inline versions of creating line items and sections. Helps to
reduce the verbosity of code if a line item is planned to be included in
a final model yet but is to be used as a variable in other locations.
- Added a `.previous()` method to shift the values in a line item.
Useful for recursive calculations where we rely on previous cells to
compute the current cell.